feat: Implement jump game algorithms
This commit adds three new files: `45.jump-game-ii.py`, `55-jump-game.py`, and `nstraighthand.py`. These files contain the implementation of jump game algorithms. The `45.jump-game-ii.py` file implements the `jump` function, which calculates the minimum number of jumps needed to reach the last index of an array. The `55-jump-game.py` file implements the `canJump` function, which determines if it is possible to reach the last index of an array. The `nstraighthand.py` file implements the `isNStraightHand` function, which checks if a given hand of cards can be divided into groups of consecutive numbers.

Ref: #456